Dashboard/robot_agent/requirements.txt
paho-mqtt
psutil